+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Features
 - add support for creating and attaching middlewares, guards and interceptors to fields and resolvers
+- **Breaking Change**: remove deprecated decorators with `GraphQL` prefix and `{ array: true }` type option
 
 ## v0.10.0
 ### Features